backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-multipart
pyjwt
cryptography
pymupdf
aiofiles
google-generativeai